test: Allow putting +0/-0 on same side of branch cut
In a computing environment where +0 and -0 are different, it's expected
that atan2(+0, -x) = +pi, atan2(-0, -x) = -pi. In a computing
environment where there's only one unsigned zero, it's expected that
atan2(0, -x) = +pi.

JavaScript (and the majority if not all of runtime environments wenyan
currently run in) has signed zeros, but currently there's no portable
way to differentiate +0 and -0 in wenyan. As a result, the math library
currently is not able to place +0 and -0 on different branches for
functions like atan2.

This commit modifies the test for atan2 so that it accepts both
atan2(-0, -x) = -pi and atan2(-0, -x) = pi.